refactored connection, added traits
