+++ src/DataMapper/Pdo/Connection/AbstractConnection.php
@@ -19,18 +19,14 @@ declare(strict_types=1);
 namespace Phalcon\DataMapper\Pdo\Connection;
 
 use BadMethodCallException;
-use Generator;
 use PDO;
 use PDOStatement;
+use Phalcon\DataMapper\Pdo\Connection\Traits\FetchTrait;
+use Phalcon\DataMapper\Pdo\Connection\Traits\YieldTrait;
 use Phalcon\DataMapper\Pdo\Exception\Exception;
 use Phalcon\DataMapper\Pdo\Profiler\ProfilerInterface;
 
-use function array_merge;
-use function call_user_func_array;
-use function current;
-use function func_get_args;
 use function get_class;
-use function implode;
 use function is_array;
 use function is_bool;
 use function is_int;
@@ -40,20 +36,23 @@ use function method_exists;
  * Provides array quoting, profiling, a new `perform()` method, new `fetch*()`
  * methods
  *
+ * @property PDO|null               $pdo
+ * @property ProfilerInterface|null $profiler
+ *
+ * @method bool        beginTransaction()
+ * @method bool        commit()
  * @method string|null errorCode()
  * @method array       errorInfo()
+ * @method mixed       getAttribute(int $attribute)
+ * @method string      lastInsertId(string $name = null)
+ * @method bool        inTransaction()
+ * @method bool        rollback()
+ * @method bool        setAttribute(int $attribute, mixed $value)
  */
 abstract class AbstractConnection
 {
-    /**
-     * @var PDO|null
-     */
-    protected ?PDO $pdo = null;
-
-    /**
-     * @var ProfilerInterface|null
-     */
-    protected ?ProfilerInterface $profiler;
+    use FetchTrait;
+    use YieldTrait;
 
     /**
      * Proxies to PDO methods created for specific drivers; in particular,
@@ -77,41 +76,11 @@ abstract class AbstractConnection
             throw new BadMethodCallException($message);
         }
 
-        return $this->pdo->{$method}(...$arguments);
-    }
+        $this->profileStart($method);
 
-    /**
-     * Begins a transaction. If the profiler is enabled, the operation will
-     * be recorded.
-     *
-     * @return bool
-     */
-    public function beginTransaction(): bool
-    {
-        $this->connect();
-        $this->profiler->start(__FUNCTION__);
+        $result = $this->pdo->{$method}(...$arguments);
 
-        $result = $this->pdo->beginTransaction();
-
-        $this->profiler->finish();
-
-        return $result;
-    }
-
-    /**
-     * Commits the existing transaction. If the profiler is enabled, the
-     * operation will be recorded.
-     *
-     * @return bool
-     */
-    public function commit(): bool
-    {
-        $this->connect();
-        $this->profiler->start(__FUNCTION__);
-
-        $result = $this->pdo->commit();
-
-        $this->profiler->finish();
+        $this->profileFinish();
 
         return $result;
     }
@@ -126,270 +95,15 @@ abstract class AbstractConnection
      */
     abstract public function disconnect(): void;
 
-    /**
-     * Executes an SQL statement and returns the number of affected rows. If
-     * the profiler is enabled, the operation will be recorded.
-     *
-     * @param string $statement
-     *
-     * @return int
-     */
-    public function exec(string $statement): int
+    public function exec(string $statement = ''): int|false
     {
-        $this->connect();
-        $this->profiler->start(__FUNCTION__);
+        $this->profileStart(__METHOD__);
 
-        $affectedRows = $this->pdo->exec($statement);
+        $result = $this->pdo->exec($statement);
 
-        $this->profiler->finish($statement);
+        $this->profileFinish($statement);
 
-        return $affectedRows;
-    }
-
-    /**
-     * Performs a statement and returns the number of affected rows.
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return int
-     * @throws Exception
-     */
-    public function fetchAffected(string $statement, array $values = []): int
-    {
-        $sth = $this->perform($statement, $values);
-
-        return $sth->rowCount();
-    }
-
-    /**
-     * Fetches a sequential array of rows from the database; the rows are
-     * returned as associative arrays.
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchAll(string $statement, array $values = []): array
-    {
-        return $this->fetchData(
-            "fetchAll",
-            [PDO::FETCH_ASSOC],
-            $statement,
-            $values
-        );
-    }
-
-    /**
-     * Fetches an associative array of rows from the database; the rows are
-     * returned as associative arrays, and the array of rows is keyed on the
-     * first column of each row.
-     *
-     * If multiple rows have the same first column value, the last row with
-     * that value will overwrite earlier rows. This method is more resource
-     * intensive and should be avoided if possible.
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchAssoc(string $statement, array $values = []): array
-    {
-        $data = [];
-        $sth  = $this->perform($statement, $values);
-
-        $row = $sth->fetch(PDO::FETCH_ASSOC);
-        while ($row) {
-            $data[current($row)] = $row;
-
-            $row = $sth->fetch(PDO::FETCH_ASSOC);
-        }
-
-        return $data;
-    }
-
-    /**
-     * Fetches a column of rows as a sequential array (default first one).
-     *
-     * @param string $statement
-     * @param array  $values
-     * @param int    $column
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchColumn(
-        string $statement,
-        array $values = [],
-        int $column = 0
-    ): array {
-        return $this->fetchData(
-            "fetchAll",
-            [PDO::FETCH_COLUMN, $column],
-            $statement,
-            $values
-        );
-    }
-
-    /**
-     * Fetches multiple from the database as an associative array. The first
-     * column will be the index key. The default flags are
-     * PDO::FETCH_ASSOC | PDO::FETCH_GROUP
-     *
-     * @param string $statement
-     * @param array  $values
-     * @param int    $flags
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchGroup(
-        string $statement,
-        array $values = [],
-        int $flags = PDO::FETCH_ASSOC
-    ): array {
-        return $this->fetchData(
-            "fetchAll",
-            [PDO::FETCH_GROUP | $flags],
-            $statement,
-            $values
-        );
-    }
-
-    /**
-     * Fetches one row from the database as an object where the column values
-     * are mapped to object properties.
-     *
-     * Since PDO injects property values before invoking the constructor, any
-     * initializations for defaults that you potentially have in your object's
-     * constructor, will override the values that have been injected by
-     * `fetchObject`. The default object returned is `\stdClass`
-     *
-     * @param string $statement
-     * @param array  $values
-     * @param string $className
-     * @param array  $arguments
-     *
-     * @return object
-     * @throws Exception
-     */
-    public function fetchObject(
-        string $statement,
-        array $values = [],
-        string $className = "stdClass",
-        array $arguments = []
-    ): object {
-        $sth = $this->perform($statement, $values);
-
-        return $sth->fetchObject($className, $arguments);
-    }
-
-    /**
-     * Fetches a sequential array of rows from the database; the rows are
-     * returned as objects where the column values are mapped to object
-     * properties.
-     *
-     * Since PDO injects property values before invoking the constructor, any
-     * initializations for defaults that you potentially have in your object's
-     * constructor, will override the values that have been injected by
-     * `fetchObject`. The default object returned is `\stdClass`
-     *
-     * @param string $statement
-     * @param array  $values
-     * @param string $class
-     * @param array  $arguments
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchObjects(
-        string $statement,
-        array $values = [],
-        string $class = "stdClass",
-        array $arguments = []
-    ): array {
-        $sth = $this->perform($statement, $values);
-
-        return $sth->fetchAll(PDO::FETCH_CLASS, $class, $arguments);
-    }
-
-    /**
-     * Fetches one row from the database as an associative array.
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchOne(string $statement, array $values = []): array
-    {
-        return $this->fetchData(
-            "fetch",
-            [PDO::FETCH_ASSOC],
-            $statement,
-            $values
-        );
-    }
-
-    /**
-     * Fetches an associative array of rows as key-value pairs (first column is
-     * the key, second column is the value).
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchPairs(string $statement, array $values = []): array
-    {
-        return $this->fetchData(
-            "fetchAll",
-            [PDO::FETCH_KEY_PAIR],
-            $statement,
-            $values
-        );
-    }
-
-    /**
-     * Fetches an associative array of rows uniquely. The rows are returned as
-     * associative arrays.
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return array
-     * @throws Exception
-     */
-    public function fetchUnique(string $statement, array $values = []): array
-    {
-        return $this->fetchData(
-            "fetchAll",
-            [PDO::FETCH_UNIQUE],
-            $statement,
-            $values
-        );
-    }
-
-    /**
-     * Fetches the very first value (i.e., first column of the first row).
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return mixed
-     * @throws Exception
-     */
-    public function fetchValue(string $statement, array $values = []): mixed
-    {
-        $sth = $this->perform($statement, $values);
-
-        return $sth->fetchColumn();
+        return $result;
     }
 
     /**
@@ -405,17 +119,11 @@ abstract class AbstractConnection
     }
 
     /**
-     * Retrieve a database connection attribute
-     *
-     * @param int $attribute
-     *
-     * @return bool|int|string|array|null
+     * Return an array of available PDO drivers
      */
-    public function getAttribute(int $attribute): bool | int | string | array | null
+    public static function getAvailableDrivers(): array
     {
-        $this->connect();
-
-        return $this->pdo->getAttribute($attribute);
+        return PDO::getAvailableDrivers();
     }
 
     /**
@@ -433,67 +141,48 @@ abstract class AbstractConnection
     /**
      * Returns the Profiler instance.
      *
-     * @return ProfilerInterface
+     * @return ProfilerInterface|null
      */
-    public function getProfiler(): ProfilerInterface
+    public function getProfiler(): ?ProfilerInterface
     {
         return $this->profiler;
     }
 
-    /**
-     * Gets the quote parameters based on the driver
-     *
-     * @param string $driver
-     *
-     * @return array
-     */
-    public function getQuoteNames(string $driver = ""): array
-    {
-        if (true === empty($driver)) {
-            $driver = $this->getDriverName();
-        }
-
-        return match ($driver) {
-            "mysql"  => [
-                "prefix"  => "`",
-                "suffix"  => "`",
-                "find"    => "`",
-                "replace" => "``",
-            ],
-            "sqlsrv" => [
-                "prefix"  => "[",
-                "suffix"  => "]",
-                "find"    => "]",
-                "replace" => "][",
-            ],
-            default  => [
-                "prefix"  => "\"",
-                "suffix"  => "\"",
-                "find"    => "\"",
-                "replace" => "\"\"",
-            ],
-        };
-    }
-
-    /**
-     * Is a transaction currently active? If the profiler is enabled, the
-     * operation will be recorded. If the profiler is enabled, the operation
-     * will be recorded.
-     *
-     * @return bool
-     */
-    public function inTransaction(): bool
-    {
-        $this->connect();
-        $this->profiler->start(__FUNCTION__);
-
-        $result = $this->pdo->inTransaction();
-
-        $this->profiler->finish();
-
-        return $result;
-    }
-
+//    /**
+//     * Gets the quote parameters based on the driver
+//     *
+//     * @param string $driver
+//     *
+//     * @return array
+//     */
+//    public function getQuoteNames(string $driver = ""): array
+//    {
+//        if (true === empty($driver)) {
+//            $driver = $this->getDriverName();
+//        }
+//
+//        return match ($driver) {
+//            "mysql"  => [
+//                "prefix"  => "`",
+//                "suffix"  => "`",
+//                "find"    => "`",
+//                "replace" => "``",
+//            ],
+//            "sqlsrv" => [
+//                "prefix"  => "[",
+//                "suffix"  => "]",
+//                "find"    => "]",
+//                "replace" => "][",
+//            ],
+//            default  => [
+//                "prefix"  => "\"",
+//                "suffix"  => "\"",
+//                "find"    => "\"",
+//                "replace" => "\"\"",
+//            ],
+//        };
+//    }
+//
     /**
      * Is the PDO connection active?
      *
@@ -505,33 +194,13 @@ abstract class AbstractConnection
     }
 
     /**
-     * Returns the last inserted autoincrement sequence value. If the profiler
-     * is enabled, the operation will be recorded.
-     *
-     * @param string|null $name
-     *
-     * @return string
-     */
-    public function lastInsertId(string $name = null): string
-    {
-        $this->connect();
-        $this->profiler->start(__FUNCTION__);
-
-        $result = $this->pdo->lastInsertId($name);
-
-        $this->profiler->finish();
-
-        return $result;
-    }
-
-    /**
      * Performs a query with bound values and returns the resulting
      * PDOStatement; array values will be passed through `quote()` and their
      * respective placeholders will be replaced in the query string. If the
      * profiler is enabled, the operation will be recorded.
      *
-     * @param string $statement
-     * @param array  $values
+     * @param string               $statement
+     * @param array<string, mixed> $values
      *
      * @return PDOStatement
      * @throws Exception
@@ -542,7 +211,7 @@ abstract class AbstractConnection
     ): PDOStatement {
         $this->connect();
 
-        $this->profiler->start(__FUNCTION__);
+        $this->profileStart(__METHOD__);
 
         $sth = $this->prepare($statement);
         foreach ($values as $name => $value) {
@@ -551,7 +220,7 @@ abstract class AbstractConnection
 
         $sth->execute();
 
-        $this->profiler->finish($statement, $values);
+        $this->profileFinish($statement, $values);
 
         return $sth;
     }
@@ -570,11 +239,11 @@ abstract class AbstractConnection
     ): PDOStatement {
         $this->connect();
 
-        $this->profiler->start(__FUNCTION__);
+        $this->profileStart(__METHOD__);
 
         $sth = $this->pdo->prepare($statement, $options);
 
-        $this->profiler->finish($sth->queryString);
+        $this->profileFinish($sth->queryString);
 
         return $sth;
     }
@@ -592,91 +261,48 @@ abstract class AbstractConnection
     public function query(
         string $statement,
         ?int $mode = null,
-        ...$arguments
+        mixed ...$arguments
     ): PDOStatement | false {
         $this->connect();
-        $this->profiler->start(__FUNCTION__);
+        $this->profileStart(__METHOD__);
 
-        $sth = call_user_func_array(
-            [
-                $this->pdo,
-                "query",
-            ],
-            func_get_args()
-        );
+        $sth = $this->pdo->query($statement, $mode, ...$arguments);
 
-        $this->profiler->finish($sth->queryString);
+        $this->profileFinish($sth->queryString);
 
         return $sth;
     }
 
-    /**
-     * Quotes a value for use in an SQL statement. This differs from
-     * `PDO::quote()` in that it will convert an array into a string of
-     * comma-separated quoted values. The default type is `PDO::PARAM_STR`
-     *
-     * @param mixed $value
-     * @param int   $type
-     *
-     * @return string The quoted value.
-     */
-    public function quote(mixed $value, int $type = PDO::PARAM_STR): string
-    {
-        $elements = [];
-
-        $this->connect();
-
-        $element = $value;
-        $quotes  = $this->getQuoteNames();
-
-        if (true !== is_array($element)) {
-            $element = (string)$element;
-
-            return $quotes["prefix"] . $element . $quotes["suffix"];
-        }
-
-        // quote array values, not keys, then combine with commas
-        foreach ($value as $key => $element) {
-            $element        = (string)$element;
-            $elements[$key] = $quotes["prefix"] . $element . $quotes["suffix"];
-        }
-
-        return implode(", ", $elements);
-    }
-
-    /**
-     * Rolls back the current transaction, and restores autocommit mode. If the
-     * profiler is enabled, the operation will be recorded.
-     *
-     * @return bool
-     */
-    public function rollBack(): bool
-    {
-        $this->connect();
-        $this->profiler->start(__FUNCTION__);
-
-        $result = $this->pdo->rollBack();
-
-        $this->profiler->finish();
-
-        return $result;
-    }
-
-    /**
-     * Set a database connection attribute
-     *
-     * @param int   $attribute
-     * @param mixed $value
-     *
-     * @return bool
-     */
-    public function setAttribute(int $attribute, mixed $value): bool
-    {
-        $this->connect();
-
-        return $this->pdo->setAttribute($attribute, $value);
-    }
-
+//    /**
+//     * Quotes a value for use in an SQL statement. This differs from
+//     * `PDO::quote()` in that it will convert an array into a string of
+//     * comma-separated quoted values. The default type is `PDO::PARAM_STR`
+//     *
+//     * @param mixed $value
+//     * @param int   $type
+//     *
+//     * @return string The quoted value.
+//     */
+//    public function quote(mixed $value, int $type = PDO::PARAM_STR): string
+//    {
+//        $this->connect();
+//
+//        if (true !== is_array($value)) {
+//            return $this->pdo->quote((string)$value, $type);
+//        }
+//
+//        // quote array values, not keys, then combine with commas
+//        $quotes = $this->getQuoteNames();
+//        $elements = array_map(
+//            fn($element) => $quotes["prefix"]
+//                . (string)$element
+//                . $quotes["suffix"],
+//            $value
+//        );
+//
+//        return implode(", ", $elements);
+//    }
+//
     /**
      * Sets the Profiler instance.
      *
@@ -692,174 +318,6 @@ abstract class AbstractConnection
     }
 
     /**
-     * Yield results using fetchAll
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return Generator
-     * @throws Exception
-     */
-    public function yieldAll(string $statement, array $values = []): Generator
-    {
-        $sth = $this->perform($statement, $values);
-        while ($row = $sth->fetch(PDO::FETCH_ASSOC)) {
-            yield $row;
-        }
-    }
-
-    /**
-     * Yield results using fetchAssoc
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return Generator
-     * @throws Exception
-     */
-    public function yieldAssoc(
-        string $statement,
-        array $values = []
-    ): Generator {
-        $sth = $this->perform($statement, $values);
-        while ($row = $sth->fetch(PDO::FETCH_ASSOC)) {
-            $key = current($row);
-            yield $key => $row;
-        }
-    }
-
-    /**
-     * Yield results using fetchColumn
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return Generator
-     * @throws Exception
-     */
-    public function yieldColumn(
-        string $statement,
-        array $values = []
-    ): Generator {
-        $sth = $this->perform($statement, $values);
-        while ($row = $sth->fetch(PDO::FETCH_NUM)) {
-            yield $row[0];
-        }
-    }
-
-    /**
-     * Yield objects where the column values are mapped to object properties.
-     *
-     * Warning: PDO "injects property-values BEFORE invoking the constructor -
-     * in other words, if your class initializes property-values to defaults
-     * in the constructor, you will be overwriting the values injected by
-     * fetchObject() !"
-     * <https://www.php.net/manual/en/pdostatement.fetchobject.php#111744>
-     *
-     * @param string $statement
-     * @param array  $values
-     * @param string $class
-     * @param array  $arguments
-     *
-     * @return Generator
-     * @throws Exception
-     */
-    public function yieldObjects(
-        string $statement,
-        array $values = [],
-        string $class = 'stdClass',
-        array $arguments = []
-    ): Generator {
-        $sth = $this->perform($statement, $values);
-
-        if (empty($arguments)) {
-            while ($instance = $sth->fetchObject($class)) {
-                yield $instance;
-            }
-        } else {
-            while ($instance = $sth->fetchObject($class, $arguments)) {
-                yield $instance;
-            }
-        }
-    }
-
-    /**
-     * Yield key-value pairs (key => value)
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return Generator
-     * @throws Exception
-     */
-    public function yieldPairs(
-        string $statement,
-        array $values = []
-    ): Generator {
-        $sth = $this->perform($statement, $values);
-        while ($row = $sth->fetch(PDO::FETCH_NUM)) {
-            yield $row[0] => $row[1];
-        }
-    }
-
-    /**
-     * Yield results using `fetchAll` and `FETCH_UNIQUE`
-     *
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return Generator
-     * @throws Exception
-     */
-    public function yieldUnique(
-        string $statement,
-        array $values = []
-    ): Generator {
-        $sth = $this->perform($statement, $values);
-        while ($row = $sth->fetch(PDO::FETCH_UNIQUE)) {
-            $key = array_shift($row);
-            yield $key => $row;
-        }
-    }
-
-    /**
-     * Helper method to get data from PDO based on the method passed
-     *
-     * @param string $method
-     * @param array  $arguments
-     * @param string $statement
-     * @param array  $values
-     *
-     * @return array
-     * @throws Exception
-     */
-    protected function fetchData(
-        string $method,
-        array $arguments,
-        string $statement,
-        array $values = []
-    ): array {
-        $sth    = $this->perform($statement, $values);
-        $result = call_user_func_array(
-            [
-                $sth,
-                $method,
-            ],
-            $arguments
-        );
-
-        /**
-         * If this returns boolean or anything other than an array, return
-         * an empty array back
-         */
-        if (true !== is_array($result)) {
-            $result = [];
-        }
-
-        return $result;
-    }
-
-    /**
      * Bind a value using the proper PDO::PARAM_* type.
      *
      * @param PDOStatement $statement
@@ -871,28 +329,45 @@ abstract class AbstractConnection
         mixed $name,
         mixed $arguments
     ): void {
-        $key = $name;
-        if (is_int($key)) {
-            $key = $key + 1;
-        }
+        $key = is_int($name) ? $name + 1 : $name;
 
-        $parameters = [$key, $arguments];
         if (is_array($arguments)) {
             $type = $arguments[1] ?? PDO::PARAM_STR;
 
-            if (PDO::PARAM_BOOL === $type && is_bool($arguments[0])) {
+            if ($type === PDO::PARAM_BOOL && is_bool($arguments[0])) {
                 $arguments[0] = $arguments[0] ? "1" : "0";
             }
 
-            $parameters = array_merge([$key], $arguments);
+            $statement->bindValue($key, $arguments[0], $type);
+        } else {
+            $statement->bindValue($key, $arguments);
         }
+    }
 
-        call_user_func_array(
-            [
-                $statement,
-                "bindValue",
-            ],
-            $parameters
-        );
+    /**
+     * @param string|null $statement
+     * @param array       $values
+     *
+     * @return void
+     */
+    protected function profileFinish(
+        string $statement = null,
+        array $values = []
+    ): void {
+        if ($this->profiler) {
+            $this->profiler->finish($statement, $values);
+        }
+    }
+
+    /**
+     * @param string $method
+     *
+     * @return void
+     */
+    protected function profileStart(string $method): void
+    {
+        if ($this->profiler) {
+            $this->profiler->start($method);
+        }
     }
 }
